P1_S4 .3
DB_Update - Column received_at erased, it was not useful

Servers - The servers can listen the port and also upload the information to the DB. The web server works perfeclty.

It was tried using the same machine but in two different instances.

What's next? Start the communication between the data bases in order to make a MASTER-MASTER architecture