P1_S4 update
The app send GPS information using TCP protocol without stop, it can be executed in background plane.

Also, the server works perfectly and the web server too. It's required to work with DNS, change line 74 of web_server file in order to work properly.

Also, it's need to make portforwarding from 5001 port to the priv IP